--- a/src/runtime/traceback.cpp
+++ b/src/runtime/traceback.cpp
@@ -96,6 +96,7 @@ Box* BoxedTraceback::getLines(Box* b) {
 
     if (!tb->py_lines) {
         BoxedList* lines = new BoxedList();
+        lines->ensure(tb->lines.size());
         for (auto line : tb->lines) {
             auto l = new BoxedTuple({ boxString(line->file), boxString(line->func), boxInt(line->line) });
             listAppendInternal(lines, l);