A few changes (mostly dealing with scanning) that help GC perf.

```
              pyston (calibration)                      :    1.0s baseline: 1.0 (-0.4%)
              pyston django_migrate.py                  :    5.9s baseline: 10.7 (-45.1%)
              pyston interp2.py                         :    6.0s baseline: 5.9 (+1.5%)
              pyston raytrace.py                        :    6.7s baseline: 6.8 (-1.1%)
              pyston nbody.py                           :    8.9s baseline: 9.3 (-4.0%)
              pyston fannkuch.py                        :    8.1s baseline: 8.0 (+0.8%)
              pyston chaos.py                           :   21.3s baseline: 20.5 (+4.2%)
              pyston fasta.py                           :    6.0s baseline: 5.6 (+8.2%)
              pyston pidigits.py                        :    5.4s baseline: 5.5 (-0.6%)
              pyston richards.py                        :    2.0s baseline: 2.0 (-0.9%)
              pyston deltablue.py                       :    2.1s baseline: 2.1 (+0.8%)
              pyston (geomean-8999)                     :    5.9s baseline: 6.2 (-5.0%)
```